requirements.txt
biopython>=1.85
numpy>=2.3
pandas>=2.3
matplotlib>=3.10
scikit-learn>=1.7
